+++ rugosa/__init__.py
@@ -1,3 +1,4 @@
 from .emulation import Emulator
+from .disassembly import *
 
 __version__ = "0.1.1"