+++ rugosa/emulation/call_hooks/stdlib/__init__.py
@@ -1,1 +1,2 @@
-from .libc import *
+
+from . import libc