+++ rugosa/emulation/call_hooks/win_api/win_constants.py
@@ -1,4 +1,6 @@
-# Constants that will be used throughout the Windows API functions
+"""
+Constants that will be used throughout the Windows API functions
+"""
 
 from enum import IntEnum, IntFlag